Separate page for contributors & About us page Optimize Fixed (#258)
* New branch

* BTT

* Revert "BTT"

This reverts commit 08433ded5d93b8124c7f031474fd9fd92939d6ac.

* Contributors

* fixed

* fixed
