+++ src/pages/About/about.css
@@ -46,73 +46,6 @@
   margin-bottom: 30px;
 }
 
-.about-container > .container-grid {
-  padding-top: 1vw;
-  width: 100%;
-  display: flex;
-  flex-wrap: wrap;
-  gap: 40px;
-}
-
-.container-grid > .MuiGrid-item-root {
-  display: flex;
-  justify-content: center;
-}
-
-.about-container .container-grid .card {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  width: 400px;
-  height: 350px;
-  border: 0.5px solid #008080;
-  transition: box-shadow 1s;
-}
-
-.about-container .container-grid .card:hover {
-  box-shadow: 10px 10px #008080;
-}
-
-.about-container .container-grid .card .cardLink .img {
-  padding: 10px;
-  border-radius: 50%;
-  width: 180px;
-  height: 180px;
-  background-size: cover !important;
-  background-repeat: no-repeat !important;
-  background-position: center !important;
-  object-fit: cover !important;
-}
-
-.about-container .container-grid .card .card-name {
-  font-size: 28px;
-  color: #008080;
-  text-align: center;
-}
-
-.container-grid .card .card-bottom {
-  display: flex;
-  justify-content: center;
-  flex-wrap: wrap;
-  gap: 5px;
-}
-
-.container-grid .card .card-bottom .text-color {
-  font-size: 1.2vw;
-  color: #008080;
-}
-
-.container-grid
-  > .MuiGrid-item-root
-  > .card
-  > .MuiCardContent-root
-  > .card-bottom
-  > .temp {
-  margin-left: 5px;
-  margin-right: 5px;
-  font-size: 30px;
-  color: #008080;
-}
 
 @media (max-width: 586px) {
   .about-container {
@@ -127,23 +60,6 @@
     margin: auto;
   }
 
-  .container-grid
-    > .MuiGrid-item-root
-    > .card
-    > .MuiCardContent-root
-    > .card-bottom
-    > .text-color {
-    font-size: 20px;
-  }
-
-  .container-grid > .MuiGrid-item-root > .card {
-    width: 100%;
-  }
-
-  .container-grid > .MuiGrid-item-root > .card > .cardLink > .img {
-    width: 100px;
-    height: 100px;
-  }
 }
 
 @media (min-width: 600px) {
@@ -160,23 +76,6 @@
     font-size: 50px;
   }
 
-  .container-grid
-    > .MuiGrid-item-root
-    > .card
-    > .MuiCardContent-root
-    > .card-bottom
-    > .text-color {
-    font-size: 24px;
-  }
-
-  .container-grid > .MuiGrid-item-root > .card {
-    width: 300px;
-  }
-
-  .container-grid > .MuiGrid-item-root > .card > .cardLink > .img {
-    width: 150px;
-    height: 150px;
-  }
 }
 
 @media (min-width: 600px) and (max-width: 1024px) {
@@ -191,22 +90,4 @@
     font-size: 40px;
   }
 
-  .container-grid
-    > .MuiGrid-item-root
-    > .card
-    > .MuiCardContent-root
-    > .card-bottom
-    > .text-color {
-    font-size: 20px;
-  }
-
-  .container-grid > .MuiGrid-item-root > .card {
-    width: 300px;
-    height: 50%;
-  }
-
-  .container-grid > .MuiGrid-item-root > .card > .cardLink > .img {
-    width: 120px;
-    height: 120px;
-  }
 }
